requirements.txt
apiai==1.2.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
future==0.18.2
futures==3.1.1
geojson==2.5.0
goslate==1.5.1
idna==2.8
numpy==1.18.2
pycparser==2.19
pyowm==2.10.0
python-telegram-bot==12.2.0
requests==2.22.0
six==1.14.0
tornado==6.0.3
urllib3==1.25.7
yandex-translater==6.0
covid-data-api==1.5
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.2.1
pyparsing==2.4.6
python-dateutil==2.8.1
